+++ Source/Krypton Toolkit Examples/KryptonFolderBrowserDialog Example/Form1.Designer.cs
@@ -1,5 +1,5 @@
 ﻿
-namespace KryptonFontDialog_Example_2019
+namespace KryptonFolderBrowserDialog_Example_2019
 {
     partial class Form1
     {
@@ -29,17 +29,11 @@ namespace KryptonFontDialog_Example_2019
         /// </summary>
         private void InitializeComponent()
         {
-            groupBoxPalette = new Krypton.Toolkit.KryptonGroupBox();
-            buttonShowFontDialog = new Krypton.Toolkit.KryptonButton();
+            buttonFileOpenDialog = new Krypton.Toolkit.KryptonButton();
             kryptonPanel1 = new Krypton.Toolkit.KryptonPanel();
+            kryptonThemeComboBox1 = new Krypton.Toolkit.KryptonThemeComboBox();
+            kryptonLabel1 = new Krypton.Toolkit.KryptonLabel();
             kryptonGroupBox2 = new Krypton.Toolkit.KryptonGroupBox();
-            chkShowEffects = new Krypton.Toolkit.KryptonCheckBox();
-            chkShowApply = new Krypton.Toolkit.KryptonCheckBox();
-            chkAllowVectorFonts = new Krypton.Toolkit.KryptonCheckBox();
-            chkAllowSimulations = new Krypton.Toolkit.KryptonCheckBox();
-            chkAllowScriptChange = new Krypton.Toolkit.KryptonCheckBox();
-            chkShowColour = new Krypton.Toolkit.KryptonCheckBox();
-            chkShowHelp = new Krypton.Toolkit.KryptonCheckBox();
             kryptonGroupBox1 = new Krypton.Toolkit.KryptonGroupBox();
             kryptonTextBox1 = new Krypton.Toolkit.KryptonTextBox();
             kryptonNumericUpDown1 = new Krypton.Toolkit.KryptonNumericUpDown();
@@ -47,131 +41,80 @@ namespace KryptonFontDialog_Example_2019
             kryptonCheckBox1 = new Krypton.Toolkit.KryptonCheckBox();
             kryptonListBox1 = new Krypton.Toolkit.KryptonListBox();
             kryptonComboBox1 = new Krypton.Toolkit.KryptonComboBox();
-            ((System.ComponentModel.ISupportInitialize)groupBoxPalette).BeginInit();
-            ((System.ComponentModel.ISupportInitialize)groupBoxPalette.Panel).BeginInit();
+            kryptonCheckBox2 = new Krypton.Toolkit.KryptonCheckBox();
             ((System.ComponentModel.ISupportInitialize)kryptonPanel1).BeginInit();
             kryptonPanel1.SuspendLayout();
+            ((System.ComponentModel.ISupportInitialize)kryptonThemeComboBox1).BeginInit();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox2).BeginInit();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox2.Panel).BeginInit();
-            kryptonGroupBox2.Panel.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox1).BeginInit();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox1.Panel).BeginInit();
             kryptonGroupBox1.Panel.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)kryptonComboBox1).BeginInit();
             SuspendLayout();
             // 
-            // groupBoxPalette
+            // buttonFileOpenDialog
             // 
-            groupBoxPalette.Location = new System.Drawing.Point(12, 12);
-            groupBoxPalette.Name = "groupBoxPalette";
-            groupBoxPalette.Size = new System.Drawing.Size(353, 135);
-            groupBoxPalette.TabIndex = 18;
-            groupBoxPalette.Values.Heading = "Palette";
-            // 
-            // buttonShowFontDialog
-            // 
-            buttonShowFontDialog.CornerRoundingRadius = -1F;
-            buttonShowFontDialog.Location = new System.Drawing.Point(264, 323);
-            buttonShowFontDialog.Name = "buttonShowFontDialog";
-            buttonShowFontDialog.Size = new System.Drawing.Size(307, 59);
-            buttonShowFontDialog.TabIndex = 19;
-            buttonShowFontDialog.UACShieldIconSize = Krypton.Toolkit.UACShieldIconSize.ExtraSmall;
-            buttonShowFontDialog.Values.Text = "Show FontDialog";
-            buttonShowFontDialog.Click += ButtonShowFontDialog_Click;
+            buttonFileOpenDialog.Anchor = System.Windows.Forms.AnchorStyles.Bottom | System.Windows.Forms.AnchorStyles.Right;
+            buttonFileOpenDialog.Location = new System.Drawing.Point(264, 404);
+            buttonFileOpenDialog.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            buttonFileOpenDialog.Name = "buttonFileOpenDialog";
+            buttonFileOpenDialog.Size = new System.Drawing.Size(307, 74);
+            buttonFileOpenDialog.TabIndex = 19;
+            buttonFileOpenDialog.Values.Text = "Folder Browser Dialog";
+            buttonFileOpenDialog.Click += ButtonShowFileDialog_Click;
             // 
             // kryptonPanel1
             // 
+            kryptonPanel1.Controls.Add(kryptonThemeComboBox1);
+            kryptonPanel1.Controls.Add(kryptonLabel1);
             kryptonPanel1.Controls.Add(kryptonGroupBox2);
             kryptonPanel1.Controls.Add(kryptonGroupBox1);
-            kryptonPanel1.Controls.Add(groupBoxPalette);
-            kryptonPanel1.Controls.Add(buttonShowFontDialog);
+            kryptonPanel1.Controls.Add(buttonFileOpenDialog);
             kryptonPanel1.Dock = System.Windows.Forms.DockStyle.Fill;
             kryptonPanel1.Location = new System.Drawing.Point(0, 0);
+            kryptonPanel1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonPanel1.Name = "kryptonPanel1";
-            kryptonPanel1.Size = new System.Drawing.Size(594, 410);
+            kryptonPanel1.Size = new System.Drawing.Size(594, 513);
             kryptonPanel1.TabIndex = 20;
+            // 
+            // kryptonThemeComboBox1
+            // 
+            kryptonThemeComboBox1.DisplayMember = "30";
+            kryptonThemeComboBox1.DropDownWidth = 339;
+            kryptonThemeComboBox1.IntegralHeight = false;
+            kryptonThemeComboBox1.Location = new System.Drawing.Point(12, 73);
+            kryptonThemeComboBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            kryptonThemeComboBox1.Name = "kryptonThemeComboBox1";
+            kryptonThemeComboBox1.Size = new System.Drawing.Size(339, 25);
+            kryptonThemeComboBox1.StateCommon.ComboBox.Content.TextH = Krypton.Toolkit.PaletteRelativeAlign.Near;
+            kryptonThemeComboBox1.TabIndex = 25;
+            kryptonThemeComboBox1.ValueMember = "30";
+            // 
+            // kryptonLabel1
+            // 
+            kryptonLabel1.LabelStyle = Krypton.Toolkit.LabelStyle.NormalPanel;
+            kryptonLabel1.Location = new System.Drawing.Point(12, 35);
+            kryptonLabel1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
+            kryptonLabel1.Name = "kryptonLabel1";
+            kryptonLabel1.Size = new System.Drawing.Size(68, 24);
+            kryptonLabel1.TabIndex = 24;
+            kryptonLabel1.Target = kryptonThemeComboBox1;
+            kryptonLabel1.Values.Text = "&Themes:";
             // 
             // kryptonGroupBox2
             // 
-            kryptonGroupBox2.Location = new System.Drawing.Point(372, 28);
+            kryptonGroupBox2.Location = new System.Drawing.Point(372, 35);
+            kryptonGroupBox2.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonGroupBox2.Name = "kryptonGroupBox2";
-            // 
-            // 
-            // 
-            kryptonGroupBox2.Panel.Controls.Add(chkShowEffects);
-            kryptonGroupBox2.Panel.Controls.Add(chkShowApply);
-            kryptonGroupBox2.Panel.Controls.Add(chkAllowVectorFonts);
-            kryptonGroupBox2.Panel.Controls.Add(chkAllowSimulations);
-            kryptonGroupBox2.Panel.Controls.Add(chkAllowScriptChange);
-            kryptonGroupBox2.Panel.Controls.Add(chkShowColour);
-            kryptonGroupBox2.Panel.Controls.Add(chkShowHelp);
-            kryptonGroupBox2.Size = new System.Drawing.Size(199, 289);
+            kryptonGroupBox2.Size = new System.Drawing.Size(199, 362);
             kryptonGroupBox2.TabIndex = 21;
             kryptonGroupBox2.Values.Heading = "Options:";
             // 
-            // chkShowEffects
-            // 
-            chkShowEffects.Checked = true;
-            chkShowEffects.CheckState = System.Windows.Forms.CheckState.Checked;
-            chkShowEffects.Location = new System.Drawing.Point(4, 180);
-            chkShowEffects.Name = "chkShowEffects";
-            chkShowEffects.Size = new System.Drawing.Size(110, 24);
-            chkShowEffects.TabIndex = 6;
-            chkShowEffects.Values.Text = "Show Effects";
-            // 
-            // chkShowApply
-            // 
-            chkShowApply.Location = new System.Drawing.Point(4, 150);
-            chkShowApply.Name = "chkShowApply";
-            chkShowApply.Size = new System.Drawing.Size(105, 24);
-            chkShowApply.TabIndex = 5;
-            chkShowApply.Values.Text = "Show Apply";
-            // 
-            // chkAllowVectorFonts
-            // 
-            chkAllowVectorFonts.Location = new System.Drawing.Point(4, 120);
-            chkAllowVectorFonts.Name = "chkAllowVectorFonts";
-            chkAllowVectorFonts.Size = new System.Drawing.Size(152, 24);
-            chkAllowVectorFonts.TabIndex = 4;
-            chkAllowVectorFonts.Values.Text = "Allow Vector Fonts";
-            // 
-            // chkAllowSimulations
-            // 
-            chkAllowSimulations.Location = new System.Drawing.Point(4, 90);
-            chkAllowSimulations.Name = "chkAllowSimulations";
-            chkAllowSimulations.Size = new System.Drawing.Size(145, 24);
-            chkAllowSimulations.TabIndex = 3;
-            chkAllowSimulations.Values.Text = "Allow Simulations";
-            // 
-            // chkAllowScriptChange
-            // 
-            chkAllowScriptChange.Location = new System.Drawing.Point(3, 60);
-            chkAllowScriptChange.Name = "chkAllowScriptChange";
-            chkAllowScriptChange.Size = new System.Drawing.Size(161, 24);
-            chkAllowScriptChange.TabIndex = 2;
-            chkAllowScriptChange.Values.Text = "Allow Script Change";
-            // 
-            // chkShowColour
-            // 
-            chkShowColour.Checked = true;
-            chkShowColour.CheckState = System.Windows.Forms.CheckState.Checked;
-            chkShowColour.Location = new System.Drawing.Point(4, 30);
-            chkShowColour.Name = "chkShowColour";
-            chkShowColour.Size = new System.Drawing.Size(111, 24);
-            chkShowColour.TabIndex = 1;
-            chkShowColour.Values.Text = "Show Colour";
-            // 
-            // chkShowHelp
-            // 
-            chkShowHelp.Location = new System.Drawing.Point(4, 0);
-            chkShowHelp.Name = "chkShowHelp";
-            chkShowHelp.Size = new System.Drawing.Size(98, 24);
-            chkShowHelp.TabIndex = 0;
-            chkShowHelp.Values.Text = "Show Help";
-            // 
             // kryptonGroupBox1
             // 
-            kryptonGroupBox1.Location = new System.Drawing.Point(12, 153);
+            kryptonGroupBox1.Location = new System.Drawing.Point(12, 191);
+            kryptonGroupBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonGroupBox1.Name = "kryptonGroupBox1";
             // 
             // 
@@ -182,13 +125,14 @@ namespace KryptonFontDialog_Example_2019
             kryptonGroupBox1.Panel.Controls.Add(kryptonCheckBox1);
             kryptonGroupBox1.Panel.Controls.Add(kryptonListBox1);
             kryptonGroupBox1.Panel.Controls.Add(kryptonComboBox1);
-            kryptonGroupBox1.Size = new System.Drawing.Size(348, 164);
+            kryptonGroupBox1.Size = new System.Drawing.Size(348, 205);
             kryptonGroupBox1.TabIndex = 20;
             kryptonGroupBox1.Values.Heading = "Style to match:";
             // 
             // kryptonTextBox1
             // 
-            kryptonTextBox1.Location = new System.Drawing.Point(156, 29);
+            kryptonTextBox1.Location = new System.Drawing.Point(156, 36);
+            kryptonTextBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonTextBox1.Name = "kryptonTextBox1";
             kryptonTextBox1.Size = new System.Drawing.Size(134, 27);
             kryptonTextBox1.TabIndex = 6;
@@ -196,24 +140,25 @@ namespace KryptonFontDialog_Example_2019
             // 
             // kryptonNumericUpDown1
             // 
-            kryptonNumericUpDown1.Location = new System.Drawing.Point(156, 97);
+            kryptonNumericUpDown1.Location = new System.Drawing.Point(156, 122);
+            kryptonNumericUpDown1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonNumericUpDown1.Name = "kryptonNumericUpDown1";
             kryptonNumericUpDown1.Size = new System.Drawing.Size(90, 26);
             kryptonNumericUpDown1.TabIndex = 5;
             // 
             // kryptonButton1
             // 
-            kryptonButton1.CornerRoundingRadius = -1F;
-            kryptonButton1.Location = new System.Drawing.Point(156, 65);
+            kryptonButton1.Location = new System.Drawing.Point(156, 82);
+            kryptonButton1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonButton1.Name = "kryptonButton1";
-            kryptonButton1.Size = new System.Drawing.Size(90, 25);
+            kryptonButton1.Size = new System.Drawing.Size(90, 31);
             kryptonButton1.TabIndex = 4;
-            kryptonButton1.UACShieldIconSize = Krypton.Toolkit.UACShieldIconSize.ExtraSmall;
             kryptonButton1.Values.Text = "kryptonButton1";
             // 
             // kryptonCheckBox1
             // 
-            kryptonCheckBox1.Location = new System.Drawing.Point(156, 4);
+            kryptonCheckBox1.Location = new System.Drawing.Point(156, 5);
+            kryptonCheckBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonCheckBox1.Name = "kryptonCheckBox1";
             kryptonCheckBox1.Size = new System.Drawing.Size(152, 24);
             kryptonCheckBox1.TabIndex = 2;
@@ -222,40 +167,54 @@ namespace KryptonFontDialog_Example_2019
             // kryptonListBox1
             // 
             kryptonListBox1.Items.AddRange(new object[] { "Font 1", "Font 2", "Font 1Font 1Font 1Font 1" });
-            kryptonListBox1.Location = new System.Drawing.Point(4, 29);
+            kryptonListBox1.Location = new System.Drawing.Point(4, 36);
+            kryptonListBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonListBox1.Name = "kryptonListBox1";
-            kryptonListBox1.Size = new System.Drawing.Size(120, 96);
+            kryptonListBox1.Padding = new System.Windows.Forms.Padding(1, 2, 1, 2);
+            kryptonListBox1.Size = new System.Drawing.Size(120, 120);
             kryptonListBox1.TabIndex = 1;
             // 
             // kryptonComboBox1
             // 
-            kryptonComboBox1.CornerRoundingRadius = -1F;
             kryptonComboBox1.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
             kryptonComboBox1.DropDownWidth = 121;
             kryptonComboBox1.IntegralHeight = false;
             kryptonComboBox1.Items.AddRange(new object[] { "Microsoft Sans Serif" });
-            kryptonComboBox1.Location = new System.Drawing.Point(4, 4);
+            kryptonComboBox1.Location = new System.Drawing.Point(4, 5);
+            kryptonComboBox1.Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             kryptonComboBox1.Name = "kryptonComboBox1";
             kryptonComboBox1.Size = new System.Drawing.Size(146, 25);
             kryptonComboBox1.StateCommon.ComboBox.Content.TextH = Krypton.Toolkit.PaletteRelativeAlign.Near;
             kryptonComboBox1.TabIndex = 0;
             kryptonComboBox1.Text = "Microsoft Sans Serif";
             // 
+            // kryptonCheckBox2
+            // 
+            kryptonCheckBox2.Checked = true;
+            kryptonCheckBox2.CheckState = System.Windows.Forms.CheckState.Checked;
+            kryptonCheckBox2.Location = new System.Drawing.Point(4, 180);
+            kryptonCheckBox2.Name = "kryptonCheckBox2";
+            kryptonCheckBox2.Size = new System.Drawing.Size(143, 24);
+            kryptonCheckBox2.TabIndex = 6;
+            kryptonCheckBox2.Values.Text = "Restore Directory";
+            // 
             // Form1
             // 
-            AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            AutoScaleDimensions = new System.Drawing.SizeF(8F, 20F);
             AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            ClientSize = new System.Drawing.Size(594, 410);
+            ClientSize = new System.Drawing.Size(594, 513);
             Controls.Add(kryptonPanel1);
+            FormTitleAlign = Krypton.Toolkit.PaletteRelativeAlign.Inherit;
+            Margin = new System.Windows.Forms.Padding(3, 4, 3, 4);
             Name = "Form1";
-            Text = "KryptonFontDialog Example";
-            ((System.ComponentModel.ISupportInitialize)groupBoxPalette.Panel).EndInit();
-            ((System.ComponentModel.ISupportInitialize)groupBoxPalette).EndInit();
+            SizeGripStyle = System.Windows.Forms.SizeGripStyle.Show;
+            Text = "KryptonFolderBrowserDialog Example";
+            Controls.SetChildIndex(kryptonPanel1, 0);
             ((System.ComponentModel.ISupportInitialize)kryptonPanel1).EndInit();
             kryptonPanel1.ResumeLayout(false);
+            kryptonPanel1.PerformLayout();
+            ((System.ComponentModel.ISupportInitialize)kryptonThemeComboBox1).EndInit();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox2.Panel).EndInit();
-            kryptonGroupBox2.Panel.ResumeLayout(false);
-            kryptonGroupBox2.Panel.PerformLayout();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox2).EndInit();
             ((System.ComponentModel.ISupportInitialize)kryptonGroupBox1.Panel).EndInit();
             kryptonGroupBox1.Panel.ResumeLayout(false);
@@ -266,25 +225,19 @@ namespace KryptonFontDialog_Example_2019
         }
 
         #endregion
-
-        private Krypton.Toolkit.KryptonGroupBox groupBoxPalette;
-        private Krypton.Toolkit.KryptonButton buttonShowFontDialog;
+        private Krypton.Toolkit.KryptonButton buttonFileOpenDialog;
         private Krypton.Toolkit.KryptonPanel kryptonPanel1;
         private Krypton.Toolkit.KryptonGroupBox kryptonGroupBox1;
-        private Krypton.Toolkit.KryptonNumericUpDown kryptonNumericUpDown1;
-        private Krypton.Toolkit.KryptonButton kryptonButton1;
-        private Krypton.Toolkit.KryptonCheckBox kryptonCheckBox1;
+        private Krypton.Toolkit.KryptonGroupBox kryptonGroupBox2;
         private Krypton.Toolkit.KryptonListBox kryptonListBox1;
         private Krypton.Toolkit.KryptonComboBox kryptonComboBox1;
+        private Krypton.Toolkit.KryptonCheckBox kryptonCheckBox1;
         private Krypton.Toolkit.KryptonTextBox kryptonTextBox1;
-        private Krypton.Toolkit.KryptonGroupBox kryptonGroupBox2;
-        private Krypton.Toolkit.KryptonCheckBox chkShowHelp;
-        private Krypton.Toolkit.KryptonCheckBox chkShowColour;
-        private Krypton.Toolkit.KryptonCheckBox chkShowEffects;
-        private Krypton.Toolkit.KryptonCheckBox chkShowApply;
-        private Krypton.Toolkit.KryptonCheckBox chkAllowVectorFonts;
-        private Krypton.Toolkit.KryptonCheckBox chkAllowSimulations;
-        private Krypton.Toolkit.KryptonCheckBox chkAllowScriptChange;
+        private Krypton.Toolkit.KryptonButton kryptonButton1;
+        private Krypton.Toolkit.KryptonNumericUpDown kryptonNumericUpDown1;
+        private Krypton.Toolkit.KryptonThemeComboBox kryptonThemeComboBox1;
+        private Krypton.Toolkit.KryptonLabel kryptonLabel1;
+        private Krypton.Toolkit.KryptonCheckBox kryptonCheckBox2;
     }
 }