+++ Source/Krypton Toolkit Examples/KryptonOpenFileDialog Example/Form1.cs
@@ -1,17 +1,13 @@
 ﻿using System;
-using System.ComponentModel;
 using System.Windows.Forms;
 
 using Krypton.Toolkit;
 
-namespace KryptonFontDialog_Example_2019
+namespace KryptonOpenFileDialog_Example_2019
 {
     public partial class Form1 : KryptonForm
     {
-        public Form1()
-        {
-            InitializeComponent();
-        }
+        public Form1() => InitializeComponent();
 
         private void ButtonShowFileDialog_Click(object sender, EventArgs e)
         {
@@ -31,12 +27,7 @@ namespace KryptonFontDialog_Example_2019
                 Multiselect = chkMultiselect.Checked,
                 ReadOnlyChecked = chkReadOnlyChecked.Checked, // Get only
                 RestoreDirectory = chkRestoreDirectory.Checked,
-                // ShowHelp = true, // TODO Ookii ??
-                // ShowPreview = true, // TODO Ookii ??
-                // ShowReadOnly = true,// TODO Ookii ??
-                // Site = ISite, // N/A
                 SupportMultiDottedExtensions = true,
-                // Tag = "",// N/A
                 Title = @"Select a File to Open",
                 ValidateNames = true,
                 Icon = this.Icon
@@ -46,11 +37,6 @@ namespace KryptonFontDialog_Example_2019
             {
                 KryptonMessageBox.Show(this, kfd.FileName, @"File chosen is");
             }
-        }
-
-        private void Form1_Resize(object sender, EventArgs e)
-        {
-
         }
     }
 }