+++ Documents/Help/Changelog.md
@@ -1,6 +1,7 @@
 # <img src="https://github.com/Krypton-Suite/Standard-Toolkit-Demos/blob/master/Krypton.png?raw=true"> Standard Toolkit Demos - ChangeLog
 
 ## 2023-04-## - Build 2304 - April 2023
+* Add `KryptonFolderBrowserDialog` example
 * Add `KryptonOpenFileDialog` example
 * Add `KryptonSaveFileDialog` example